+++ openquake/hazardlib/source/base.py
@@ -257,8 +257,7 @@ class ParametricSeismicSource(with_metaclass(abc.ABCMeta, BaseSeismicSource)):
     def __init__(self, source_id, name, tectonic_region_type, mfd,
                  rupture_mesh_spacing, magnitude_scaling_relationship,
                  rupture_aspect_ratio, temporal_occurrence_model):
-        super(ParametricSeismicSource, self). \
-            __init__(source_id, name, tectonic_region_type)
+        super().__init__(source_id, name, tectonic_region_type)
 
         if rupture_mesh_spacing is not None and not rupture_mesh_spacing > 0:
             raise ValueError('rupture mesh spacing must be positive')
