+++ openquake/hazardlib/source/point.py
@@ -22,7 +22,7 @@ from openquake.hazardlib.geo import Point, geodetic
 from openquake.hazardlib.geo.surface.planar import PlanarSurface
 from openquake.hazardlib.source.base import ParametricSeismicSource
 from openquake.hazardlib.source.rupture import ParametricProbabilisticRupture
-from openquake.hazardlib.calc.filters import angular_distance, KM_TO_DEGREES
+from openquake.hazardlib.geo.utils import angular_distance, KM_TO_DEGREES
 
 
 @with_slots
